Restructuring first pass
